Edit file name property_test to test_property, formatting for Property class changed to match database format/ordering
